+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/common/client/StorefrontApiRequestExecutor.kt
@@ -11,7 +11,7 @@ import com.shopify.buy3.Storefront.QueryRootQuery
 import timber.log.Timber
 import java.security.MessageDigest
 
-class StorefrontRequestExecutor(
+class StorefrontApiRequestExecutor(
     private val client: GraphClient,
     private val lruCache: LruCache<String, GraphCallResult.Success<Storefront.QueryRoot>>,
 ) {
