+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/common/di/AppModule.kt
@@ -20,7 +20,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkout_sdk_mobile_buy_integration_sample.di
+package com.shopify.checkout_sdk_mobile_buy_integration_sample.common.di
 
 import android.app.Application
 import android.util.LruCache
@@ -29,22 +29,22 @@ import com.shopify.buy3.GraphCallResult
 import com.shopify.buy3.GraphClient
 import com.shopify.buy3.Storefront
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.BuildConfig
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.CartRepository
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.CartStorefrontApiClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.CartViewModel
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontRequestExecutor
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.data.CartRepository
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.cart.data.source.network.CartStorefrontApiClient
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontApiRequestExecutor
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.LogDatabase
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.Logger
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.logs.MIGRATION_1_2
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.home.HomeViewModel
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.logs.LogsViewModel
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.ProductsViewModel
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.CollectionRepository
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.CollectionViewModel
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.CollectionsStorefrontApiClient
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.ProductRepository
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.data.CollectionRepository
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.data.source.network.CollectionsStorefrontApiClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.ProductViewModel
-import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.ProductsStorefrontApiClient
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.data.ProductRepository
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.product.data.source.network.ProductsStorefrontApiClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.PreferencesManager
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.SettingsViewModel
 import kotlinx.coroutines.CoroutineScope
@@ -71,7 +71,7 @@ val appModules = module {
     single {
         val maxCacheEntries = 100
 
-        StorefrontRequestExecutor(
+        StorefrontApiRequestExecutor(
             lruCache = LruCache<String, GraphCallResult.Success<Storefront.QueryRoot>>(maxCacheEntries),
             client = GraphClient.build(
                 context = get(),
