+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/products/collection/CollectionViewModel.kt
@@ -26,6 +26,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.navigation.Screen
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.data.Collection
+import com.shopify.checkout_sdk_mobile_buy_integration_sample.products.collection.data.CollectionRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -66,5 +68,5 @@ class CollectionViewModel(
 sealed class CollectionUIState {
     data object Loading : CollectionUIState()
     data class Error(val error: String) : CollectionUIState()
-    data class Loaded(val collection: UICollection) : CollectionUIState()
+    data class Loaded(val collection: Collection) : CollectionUIState()
 }
